security: fix audit findings (round 1)
Addresses issues from adversarial and architecture audits:

- #118: Add agent_id filter to _mark_synced() for defense-in-depth
- #127: Add validate_table_name() call in _migrate_schema() get_columns()
- #125: Extract duplicate sanitize functions to shared kernle.mcp.sanitize module

Also closes:
- #116: agent_id already in SERVER_CONTROLLED_FIELDS (verified in kernle-cloud)
- #117: forgotten fields already in SERVER_CONTROLLED_FIELDS (verified in kernle-cloud)
